Fixed initialization issues
Tackled issue #1.

Fixed dockerfile to get it to run. Removed the line that copied package-lock.json to the docker instance. This is because it is unecessary, as long as the package.json is copied to the instance.
Fixed conflicts in the package.json. Ran npm audit fix --force in order to fix the issues from a module that was not found.