Updated package.json with correct modules
Updated the modules in the package.json, and reverted the modules hat weren't working with other modules well. Because of that I was able to remove the legacy peer deps flag in the dockerfile.

Took the time to update the packages for the frontend testing suite and made sure they all work together properly