Fixing Frontend Testing Issues
I started fixing some problems with the frontend testing files. Looking into the RecipeInformation test, I found that many were deprecated or duplicated. In order to fix this, I deprecated the old file and created a new one. There I am determining which tests I should keep.

Additionally, I found that after the prior team had added in the themes, it broke many of the tests. I am now starting to Added the ThemeProvider to the tests to get them to work again.