Fixed the python file so it wouldn't fail on startup
To fix this file, I changed the sqlite command to use tbl_name rather than tableName, since tableName doesn't exist as a column.
Change the db file to use the cookbook db in the dob folder, rather than the one in the api folder.
Added sqlite to the python container.