Updated frontend user auth logic
Added ability for login/signup to take intended route. If successful, then the page the user was trying to get to would load instead of the default profile page.

Addded Error message to auth for login and signup page.

Updated routes to match improved api.

Added logout funciton to navbar. Also added changes to the account nav dropdown depending on whether a user is signed in or not.